Make Semaphore.rawlink() start notifiers if needed.
Fixes #1287

Share the link protocol between Semaphore/Event/AsyncResult. There's some ickiness here with Cython, this might be better refactored to use composition?
