+++ CHANGES.rst
@@ -23,6 +23,10 @@
   use it in a ``with`` block to avoid leaking resources. See
   :pr:`1290`, provided by Josh Snyder.
 
+- Fix semaphores to immediately notify links if they are ready and
+  ``rawlink()`` is called. This behaves like ``Event`` and
+  ``AsyncEvent``. See :issue:`1287`, reported by Dan Milon.
+
 
 1.3.7 (2018-10-12)
 ==================
